+++ docs/conf.py
@@ -51,7 +51,7 @@ master_doc = "index"
 
 # General information about the project.
 project = "marisa-trie"
-copyright = "2016-2025, Mikhail Korobov & contributors"
+copyright = "2016-2026, Mikhail Korobov & contributors"
 author = "Mikhail Korobov & contributors"
 
 # The version info for the project you're documenting, acts as replacement for
@@ -59,7 +59,7 @@ author = "Mikhail Korobov & contributors"
 # built documents.
 #
 # The short X.Y version.
-version = "1.3.2-dev"
+version = "1.4.0"
 # The full version, including alpha/beta/rc tags.
 release = version